+++ packages/bs4/src/binders/index.ts
@@ -2,5 +2,6 @@ export { collapseOnUrlBinder } from './collapse-on-url.binder';
 export { collapseBinder } from './collapse.binder';
 export { dropdownBinder } from './dropdown.binder';
 export { expanOnUrlBinder } from './expan-on-url.binder';
-export { scrollspyStarBinder } from './scrollspy-star.binder';
+export { scrollspyClassBinder } from './scrollspy-class.binder';
 export { tooltipBinder } from './tooltip.binder';
+export { scrollToOnEventBinder } from './scroll-to-on-event.binder';