+++ packages/bs4/src/components/index.ts
@@ -1,3 +1,4 @@
+export { Bs4ScrollspyComponent } from './bs4-scrollspy/bs4-scrollspy.component';
 export { TabsComponent } from './tabs/tabs.component';
 export { DropdownComponent } from './dropdown/dropdown.component';
 export { NavbarComponent } from './navbar/navbar.component';